feat: remove exam by id

+++ src/services/ExamsServices/index.ts
@@ -1,6 +1,7 @@
 import { CreateExamDTO } from '@src/dtos/ExamsDTOS';
 import { Exam, ExamStatus } from '@src/entities/Exam';
 import { ExamsRepositories } from '@src/repositories/ExamsRepositories';
+import { NotFoundError } from '@src/utils/errors/NotFoundError';
 import { MakePartial } from '@src/utils/generics/MakePartial';
 import { logger } from '@src/utils/logger';
 
@@ -12,22 +13,35 @@ export class ExamsServices {
     type,
     status = ExamStatus.ACTIVE,
   }: MakePartial<CreateExamDTO, 'status'>): Promise<Exam> {
-    const newLaboratory = await this.examsRepositories.create({
+    const newExam = await this.examsRepositories.create({
       name,
       type,
       status,
     });
 
-    logger.debug(`Exam created: ${newLaboratory}`);
+    logger.debug(`Exam created: ${newExam}`);
 
-    return newLaboratory;
+    return newExam;
   }
 
   async findAll(): Promise<Exam[]> {
     return await this.examsRepositories.findAll();
   }
 
+  async findById(id: string): Promise<Exam | null> {
+    const exam = await this.examsRepositories.findById(id);
+
+    return exam;
+  }
+
   async deleteById(id: string): Promise<Exam> {
-    return await this.examsRepositories.deleteById(id);
+    const thisExamExists = await this.findById(id);
+
+    if (!thisExamExists) {
+      throw new NotFoundError('Exam not found.');
+    }
+
+    const removedExam = await this.examsRepositories.deleteById(id);
+    return removedExam;
   }
 }
